+++ Sources/TuistKit/Services/Cache/CacheWarmService.swift
@@ -5,13 +5,18 @@ import TuistCache
 import TuistCore
 import TuistGraph
 import TuistLoader
+import TuistPlugin
 import TuistSupport
 
 final class CacheWarmService {
     private let configLoader: ConfigLoading
+    private let manifestLoader: ManifestLoading
+    private let pluginService: PluginServicing
 
     init() {
         configLoader = ConfigLoader(manifestLoader: ManifestLoader())
+        manifestLoader = ManifestLoader()
+        pluginService = PluginService()
     }
 
     func run(path: String?, profile: String?, xcframeworks: Bool, targets: Set<String>, dependenciesOnly: Bool) throws {
@@ -31,7 +36,7 @@ final class CacheWarmService {
             config: config,
             path: path,
             cacheProfile: profile,
-            includedTargets: targets,
+            includedTargets: targets.isEmpty ? try projectTargets(at: path, config: config) : targets,
             dependenciesOnly: dependenciesOnly
         )
     }
@@ -70,5 +75,18 @@ final class CacheWarmService {
             bundleArtifactBuilder: bundleBuilder,
             contentHasher: contentHasher
         )
+    }
+
+    private func projectTargets(at path: AbsolutePath, config: Config) throws -> Set<String> {
+        let plugins = try pluginService.loadPlugins(using: config)
+        try manifestLoader.register(plugins: plugins)
+        let projects: [AbsolutePath]
+        if let workspace = try? manifestLoader.loadWorkspace(at: path) {
+            projects = workspace.projects.map { AbsolutePath(path, .init($0.pathString)) }
+        } else {
+            projects = [path]
+        }
+
+        return try Set(projects.flatMap { try manifestLoader.loadProject(at: $0).targets.map(\.name) })
     }
 }