+++ projects/fourier/test/fourier/services/focus_test.rb
@@ -7,6 +7,12 @@ module Fourier
     class FocusTest < TestCase
       def test_call
         # Given
+        Utilities::System
+          .expects(:tuist)
+          .with("cache", "warm", "--dependencies-only")
+        Utilities::System
+          .expects(:tuist)
+          .with("dependencies", "fetch")
         target = "TuistSupport"
         Utilities::System
           .expects(:tuist)